+++ src/main/java/com/oracle/dragon/util/exception/ConfigurationOCIDNotStartingByExpectedPatternException.java
@@ -1,0 +1,7 @@
+package com.oracle.dragon.util.exception;
+
+public class ConfigurationOCIDNotStartingByExpectedPatternException extends DSException {
+    public ConfigurationOCIDNotStartingByExpectedPatternException(String parameter, String expectedStartingPattern) {
+        super(ErrorCode.ConfigurationOCIDNotStartingByExpectedPattern, String.format("The configuration parameter \"%s\" doesn't start by the expected pattern \"%s\"", parameter, expectedStartingPattern));
+    }
+}